Fixup size to handle strings >2GB

+++ src/memcpy.c
@@ -5,7 +5,7 @@ void *memcpy(void *dest, const void *src, size_t n) {
 }
 
 char *strncpy(char *dest, const char *src, size_t n) {
-    int srcn = strlen(src) + 1;
+    size_t srcn = strlen(src) + 1;
     if(n > srcn) {
         memset(dest+srcn, 0, n-srcn);
         n = srcn;
@@ -14,6 +14,6 @@ char *strncpy(char *dest, const char *src, size_t n) {
 }
 
 char *strcpy(char *dest, const char *src) {
-    int n = strlen(src) + 1;
+    size_t n = strlen(src) + 1;
     return memmove(dest, src, n);
 }
